347 Add ability to specify ACL in a concise way (PR 1212)
First commit of PR 1212: Use `settings.toml` from `assets/` during development

Before, it was tricky to even test having a settings file in dev mode.
But we don't want to include it in production mode (though this is not
really a problem anymore since we use CI to build production builds,
where the settings.toml is never present).

2nd commit of PR 1212: Add ability to specify `upload.acl` concisely via object

Before, one had to always specify an XML blob, which is suuuper verbose.
With this, Studio also allows a more concise representation that's also
suitable to be included in the URL. While each string in the object is
also treated as a mustache template, with the object notation, it is not
possible to conditionally add entries. At least not yet.